bad file descriptor closed heredoc

+++ srcs/execution/redirections_utils.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   redirections_utils.c                               :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: nradal <nradal@student.42.fr>              +#+  +:+       +#+        */
+/*   By: atchougo <atchougo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/11 14:31:00 by nradal            #+#    #+#             */
-/*   Updated: 2023/02/17 10:27:33 by nradal           ###   ########.fr       */
+/*   Updated: 2023/02/28 19:05:08 by atchougo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,21 +15,30 @@
 int	redirections_closer(t_cmd_array *cmd, t_fd_saver fd_saver)
 {
 	int	i;
+	int j;
 
+	j = 0;
 	i = 0;
 	while (cmd->redir_size > i)
 	{
-		if (cmd->redir_array[i].type == e_heredoc)
+		if (j == 0 && cmd->redir_array[i].type == e_heredoc)
 		{
 			unlink("/tmp/heredoc_file");
-		}
-		// if (access(cmd->redir_array[i].file_name, W_OK) == 0)
-		// {
-			if (close(cmd->redir_array[i].file_fd) == -1)
+			if (cmd->redir_array[i].file_fd != -1 \
+				&& close(cmd->redir_array[i].file_fd) == -1)
 			{
 				perror("close");
 				return (0);
 			}
+			j++;
+		}
+		else if (cmd->redir_array[i].file_fd != -1 \
+			&& close(cmd->redir_array[i].file_fd) == -1)
+			return (0);
+		// if (access(cmd->redir_array[i].file_name, W_OK) == 0)
+		// {
+			// if (close(cmd->redir_array[i].file_fd) == -1)
+				// return (0);
 		// }
 		i++;
 	}
